Work on spectral profiler: 1.  Made Profiler colors customizable and persistent. 2.  Made Profiler preferences persistent.  Fixed a reset bug with     the cancel button. 3.  Fixed bug where profiler was not resizing. 4.  Added customizable and persistent legend settings to the profiler. 5.  Work on spectral line fitting thresholding.  Completion of work is     waiting for qwt update. 6.  Preliminary work on splatalogue spectral line lookup.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@20259 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/display/QtPlotter/QtPlotSettings.cc
@@ -62,11 +62,11 @@ void QtPlotSettings::scroll(int dx, int dy)
 void QtPlotSettings::zoomOut( double zoomFactor ){
 	for ( int i = 0; i < END_AXIS_INDEX; i++ ){
 		AxisIndex axisIndex = static_cast<AxisIndex>(i);
-		int prevSpanX = spanX(axisIndex);
+		double prevSpanX = spanX(axisIndex);
 		minX[i] = minX[i] - zoomFactor * prevSpanX;
 		maxX[i] = maxX[i] + zoomFactor * prevSpanX;
 	}
-	int prevSpanY = spanY();
+	double prevSpanY = spanY();
 	minY = minY - zoomFactor * prevSpanY;
 	maxY = maxY + zoomFactor * prevSpanY;
 	adjust();
@@ -75,11 +75,11 @@ void QtPlotSettings::zoomOut( double zoomFactor ){
 void QtPlotSettings::zoomIn( double zoomFactor ){
 	for ( int i = 0; i < END_AXIS_INDEX; i++ ){
 		AxisIndex axisIndex = static_cast<AxisIndex>(i);
-		int prevSpanX = spanX( axisIndex );
+		double prevSpanX = spanX( axisIndex );
 		minX[i] = minX[i] + zoomFactor * prevSpanX;
 		maxX[i] = maxX[i] - zoomFactor * prevSpanX;
 	}
-	int prevSpanY = spanY();
+	double prevSpanY = spanY();
 	minY = minY + zoomFactor * prevSpanY;
 	maxY = maxY - zoomFactor * prevSpanY;
 	adjust();
